+++ KeyboardKit/KeyCommandCreation.swift
@@ -3,15 +3,11 @@
 import UIKit
 
 extension UIKeyCommand {
-    /// Consistent API whether there is a title or not and on iOS 12 and 13+.
+    /// Consistent API whether there is a title or not.
     /// Tuple means usage reads more nicely too.
     convenience init(_ keys: (modifierFlags: UIKeyModifierFlags, input: String), action: Selector, title: String? = nil, wantsPriorityOverSystemBehavior: Bool = false, allowsAutomaticLocalization: Bool = true, allowsAutomaticMirroring: Bool = true) {
         if let title = title {
-            if #available(iOS 13, *) {
-                self.init(title: title, action: action, input: keys.input, modifierFlags: keys.modifierFlags)
-            } else {
-                self.init(input: keys.input, modifierFlags: keys.modifierFlags, action: action, discoverabilityTitle: title)
-            }
+            self.init(title: title, action: action, input: keys.input, modifierFlags: keys.modifierFlags)
         } else {
             self.init(input: keys.input, modifierFlags: keys.modifierFlags, action: action)
         }