+++ KeyboardKit/KeyboardWindow.swift
@@ -57,16 +57,16 @@ open class KeyboardWindow: UIWindow {
         let presentationController = topmost.presentationController!
 
         guard
-            topmost.isModal == false,
-            delegateSaysPresentationControllerShouldDismiss(presentationController)
+            topmost.isModalInPresentation == false,
+            presentationController.delegate?.presentationControllerShouldDismiss?(presentationController) ?? true
         else {
-            tellDelegatePresentationControllerDidAttemptToDismiss(presentationController)
+            presentationController.delegate?.presentationControllerDidAttemptToDismiss?(presentationController)
             return
         }
 
-        tellDelegatePresentationControllerWillDismiss(presentationController)
+        presentationController.delegate?.presentationControllerWillDismiss?(presentationController)
         topmost.presentingViewController!.dismiss(animated: true) {
-            tellDelegatePresentationControllerDidDismiss(presentationController)
+            presentationController.delegate?.presentationControllerDidDismiss?(presentationController)
         }
     }
 }
@@ -84,80 +84,16 @@ private extension UIWindow {
     }
 }
 
-private func delegateSaysPresentationControllerShouldDismiss(_ presentationController: UIPresentationController) -> Bool {
-    // TODO: Verify this matches what UIKit does. Not yet documented so should find by experimentation.
-
-    if #available(iOS 13, *), let should = presentationController.delegate?.presentationControllerShouldDismiss?(presentationController) {
-        return should
-    }
-
-    // Since Catalyst did not start until iOS 13 this warns even though the deployment target is iOS 12.
-    #if !targetEnvironment(macCatalyst)
-    if
-        let popoverPresentationController = presentationController as? UIPopoverPresentationController,
-        let should = popoverPresentationController.delegate?.popoverPresentationControllerShouldDismissPopover?(popoverPresentationController)
-    {
-        return should
-    }
-    #endif
-
-    return true
-}
-
-private func tellDelegatePresentationControllerDidAttemptToDismiss(_ presentationController: UIPresentationController) {
-    if #available(iOS 13, *) {
-        presentationController.delegate?.presentationControllerDidAttemptToDismiss?(presentationController)
-    }
-}
-
-private func tellDelegatePresentationControllerWillDismiss(_ presentationController: UIPresentationController) {
-    if #available(iOS 13, *) {
-        presentationController.delegate?.presentationControllerWillDismiss?(presentationController)
-    }
-}
-
-private func tellDelegatePresentationControllerDidDismiss(_ presentationController: UIPresentationController) {
-    // TODO: Verify whether UIKit calls both if both are implemented or whether it stops after the first one is implemented.
-
-    if #available(iOS 13, *) {
-        presentationController.delegate?.presentationControllerDidDismiss?(presentationController)
-    }
-
-    // Since Catalyst did not start until iOS 13 this warns even though the deployment target is iOS 12.
-    #if !targetEnvironment(macCatalyst)
-    if let popoverPresentationController = presentationController as? UIPopoverPresentationController {
-        popoverPresentationController.delegate?.popoverPresentationControllerDidDismissPopover?(popoverPresentationController)
-    }
-    #endif
-}
-
 private extension UIModalPresentationStyle {
     /// Whether the style itself allows the user to dismiss the presented view controller.
     var isDismissibleWithoutConfirmation: Bool {
         switch self {
         case .automatic:
             preconditionFailure("UIKit should have resolved automatic to a concrete style.")
-        case .popover:
+        case .popover, .pageSheet, .formSheet:
             return true
-        case .pageSheet, .formSheet:
-            if #available(iOS 13, *) {
-                return true
-            } else {
-                return false
-            }
         case .fullScreen, .currentContext, .custom, .overFullScreen, .overCurrentContext, .none: fallthrough @unknown default:
             return false
-        }
-    }
-}
-
-private extension UIViewController {
-    /// Same as `isModalInPresentation` on iOS 13 and later, or `isModalInPopover` on earlier versions.
-    var isModal: Bool {
-        if #available(iOS 13, *) {
-            return isModalInPresentation
-        } else {
-            return isModalInPopover
         }
     }
 }